Fix compilation errors for cross-platform builds
- Fixed Cargo.toml platform-specific eBPF dependencies
- Changed RiskConfig float types from f32 to f64 for config crate compatibility
- Added Serialize/Deserialize derives to IntelState, CveMetadata, CveRiskAssessment
- Fixed allowed_syscalls iteration in enforcement.rs
- Fixed HashMap/BTreeMap array type issues in enforcement.rs and risk_engine.rs
- Added stub modules (event_consumer_stub, tc_enforcer_stub) for non-eBPF builds
- Updated main.rs with proper cfg gating for eBPF modules
- Made remediator protobuf code conditional on feature flag
- Updated build.rs to handle missing protoc gracefully

+++ scanner-agent/src/config.rs
@@ -33,8 +33,8 @@ pub struct IntelConfig {
 
 #[derive(Debug, Clone, Deserialize)]
 pub struct RiskConfig {
-    pub minimum_cvss: f32,
-    pub minimum_epss: f32,
+    pub minimum_cvss: f64,
+    pub minimum_epss: f64,
 }
 
 #[derive(Debug, Clone, Deserialize)]
